+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Create `Autodesk.Forge` internal project.
 ### Autodesk.Forge
 - Copy `BucketsApi`, `ObjectsApi`, `TwoLeggedApi` from the original `forge-api-dotnet-client` repository.
-- Update deprecated `x-ads-region` to `region` when creating a bucket. [buckets-POST](https://aps.autodesk.com/en/docs/data/v2/reference/http/buckets-POST/)
+- Update deprecated `x-ads-region` to `region` when creating a bucket. [buckets-POST](https://aps.autodesk.com/en/docs/data/v2/reference/http/buckets-POST/) (Fix: #10)
 ### Updated
 - Update `CreateBucketAsync` to work with `region`.
 ### Tests